More updates
Moved EXTI handler location
