+++ firmware/Core/Inc/gps/gps_pps_interrupt.h
@@ -4,10 +4,6 @@
 
 #include "stm32l4xx_hal.h"
 
-extern uint32_t GPS_pps_counter;
-
-void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin);
-
 void GPIO_EXTI_Init(void);
 
 void Disable_GPIO_EXTI(void);
